requirements.txt
pydantic~=2.6.4
pydantic_core~=2.16.3
fastapi~=0.110.0
PyJWT~=2.8.0
requests~=2.31.0
SQLAlchemy~=2.0.29
alembic~=1.13.1
uvicorn~=0.29.0
python-jose==3.3.0
pydantic-settings==2.2.1
jinja2==3.1.3